+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/UpdatesPoller.kt
@@ -1,0 +1,8 @@
+package com.github.insanusmokrassar.TelegramBotAPI.bot
+
+import kotlinx.coroutines.*
+import kotlinx.io.core.Closeable
+
+interface UpdatesPoller : Closeable {
+    fun start(scope: CoroutineScope = CoroutineScope(Dispatchers.Default))
+}